- This is the 1st commit message:
Smooth going until one major problem:

In order to swap nodes, we need to know how big they are.
We have the jump targets down pat, but knowing how big the current
node is is not easy. Because it's not obvious what the current node
is, only where its start position is.

- This is the commit message #2:

Iteration. Not working, but code structure is there

- This is the commit message #3:

Continuing iteration

Currently it's looking at lines. It's not working fully,
          something wrong in the swap logic. But it's looking at lines.
          It might be nice to do this by nodes, so instead of
          highest_coincident starting on the same line, looking at
          the highest coincident on the same line and col. And
          swapping like that as well.

- This is the commit message #4:

Pilfer ts_util's swap_nodes