More informative output for `make test`
Now it stops on tests that fail. This way it's much easier to find what
failed when running the whole test suite. Previously it just kept going
and threw an exit code, but gave no summary of what failed.

The tradeoff is that the tests now need to run in sequence rather than
in parallel. It's noticeably slower, but only from like 1/3rd of a
second to 1 second. I think it's still pretty quick.

Probably should just be using busted proper anyways.