remove todo
I disagree with this now. I like the limited highlight. I've tried it
for a while with the full block and while at first it seemed cool, I
eventually found myself shrinking the duration down and down, and
eventually I just wanted it to be off. This highlight feels fine to
me, so I'm leaving it this way.