Update missing parser error string
The [] look cool and feel edgy, this makes me feel slightly more like
I'm in the movie the matrix, which is like 9/10ths of why I got into
programming in the first place. I think that the ultimate expression
of programming is programming the matrix itself. To understand the
construct in a unique way. To have the power to play with the fabric
of shit.

That's why I added the [].