Implement markdown support
Markdown represents essentially an entirely new plugin.
But it does belong in Treewalker.
So now Treewalker is just bigger. Holy cow.

This is a lot of code, and most of it was written by vibes.
Weird. The code is not quite up to snuff, but it does work.
I've put a lot of time into trying to get the code nicer - it
started as essentially one super long function. I hope it's alright now.
Weird world.